Move step related models from projects
